requirements.txt
Flask==2.3.2
bcrypt>=4.0.0
Flask-Limiter>=3.5.0
openpyxl==3.1.2
Pillow>=10.3.0
pillow-heif>=0.18.0
Werkzeug==2.3.6
gunicorn==20.1.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
reportlab>=4.0
pypdf>=4.0
PyMySQL>=1.1.0
python-dateutil>=2.8.0
APScheduler>=3.10.0
solapi>=5.0.0
